# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileJobTypesItem = typeof ProfileJobTypesItem[keyof typeof ProfileJobTypesItem];


export const ProfileJobTypesItem = {
  'full-time': 'full-time',
  remote: 'remote',
  hybrid: 'hybrid',
  contract: 'contract',
} as const;

export interface Profile {
  id: number;
  userId: string;
  skills: string[];
  yearsOfExperience: number;
  preferredLocations: string[];
  noticePeriod: string;
  jobTypes: ProfileJobTypesItem[];
  includeKeywords: string[];
  excludeKeywords: string[];
  /** @nullable */
  expectedSalaryMin: number | null;
  /** @nullable */
  expectedSalaryMax: number | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type ProfileInputJobTypesItem = typeof ProfileInputJobTypesItem[keyof typeof ProfileInputJobTypesItem];


export const ProfileInputJobTypesItem = {
  'full-time': 'full-time',
  remote: 'remote',
  hybrid: 'hybrid',
  contract: 'contract',
} as const;

export interface ProfileInput {
  /** @minItems 1 */
  skills: string[];
  /** @minimum 0 */
  yearsOfExperience: number;
  /** @minItems 1 */
  preferredLocations: string[];
  noticePeriod: string;
  /** @minItems 1 */
  jobTypes: ProfileInputJobTypesItem[];
  includeKeywords: string[];
  excludeKeywords: string[];
  /** @nullable */
  expectedSalaryMin?: number | null;
  /** @nullable */
  expectedSalaryMax?: number | null;
}

export interface ProfileUpdate {
  skills?: string[];
  /** @minimum 0 */
  yearsOfExperience?: number;
  preferredLocations?: string[];
  noticePeriod?: string;
  jobTypes?: string[];
  includeKeywords?: string[];
  excludeKeywords?: string[];
  /** @nullable */
  expectedSalaryMin?: number | null;
  /** @nullable */
  expectedSalaryMax?: number | null;
  isActive?: boolean;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  new: 'new',
  interested: 'interested',
  applied: 'applied',
  not_relevant: 'not_relevant',
  duplicate: 'duplicate',
} as const;

export interface Job {
  id: number;
  userId: string;
  title: string;
  company: string;
  location: string;
  source: string;
  applyUrl: string;
  /** @nullable */
  postedAt: string | null;
  /** AI match score 0-100 */
  matchScore: number;
  /** AI explanation of why this job matches the user profile */
  matchReason: string;
  status: JobStatus;
  skills: string[];
  /** @nullable */
  experienceRequired?: string | null;
  /** @nullable */
  salaryRange?: string | null;
  /** @nullable */
  jobType?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}

export type JobStatusUpdateStatus = typeof JobStatusUpdateStatus[keyof typeof JobStatusUpdateStatus];


export const JobStatusUpdateStatus = {
  new: 'new',
  interested: 'interested',
  applied: 'applied',
  not_relevant: 'not_relevant',
  duplicate: 'duplicate',
} as const;

export interface JobStatusUpdate {
  status: JobStatusUpdateStatus;
}

export interface Feedback {
  id: number;
  userId: string;
  jobId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface FeedbackInput {
  jobId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface TelegramStatus {
  connected: boolean;
  /** @nullable */
  chatId?: string | null;
  /** @nullable */
  username?: string | null;
}

export interface TelegramConnectInput {
  chatId: string;
  username?: string;
}

export interface DashboardStats {
  totalJobs: number;
  newJobs: number;
  interestedJobs: number;
  appliedJobs: number;
  avgMatchScore: number;
  /** @nullable */
  lastSearchAt: string | null;
  profileComplete: boolean;
  telegramConnected: boolean;
}

export interface SourceCount {
  source: string;
  count: number;
}

export interface ScoreBucket {
  range: string;
  count: number;
}

export type ListJobsParams = {
status?: ListJobsStatus;
minScore?: number;
source?: string;
location?: string;
page?: number;
limit?: number;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  new: 'new',
  interested: 'interested',
  applied: 'applied',
  not_relevant: 'not_relevant',
  duplicate: 'duplicate',
} as const;