# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalJobs: number;
  newJobs: number;
  interestedJobs: number;
  appliedJobs: number;
  avgMatchScore: number;
  /** @nullable */
  lastSearchAt: string | null;
  profileComplete: boolean;
  telegramConnected: boolean;
}