# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  userId: string;
  jobId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}