# lib/api-zod/src/generated/types/jobListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}