# lib/api-zod/src/generated/types/jobStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusUpdateStatus } from './jobStatusUpdateStatus';

export interface JobStatusUpdate {
  status: JobStatusUpdateStatus;
}