# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileJobTypesItem } from './profileJobTypesItem';

export interface Profile {
  id: number;
  userId: string;
  skills: string[];
  yearsOfExperience: number;
  preferredLocations: string[];
  noticePeriod: string;
  jobTypes: ProfileJobTypesItem[];
  includeKeywords: string[];
  excludeKeywords: string[];
  /** @nullable */
  expectedSalaryMin: number | null;
  /** @nullable */
  expectedSalaryMax: number | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}