# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  skills?: string[];
  /** @minimum 0 */
  yearsOfExperience?: number;
  preferredLocations?: string[];
  noticePeriod?: string;
  jobTypes?: string[];
  includeKeywords?: string[];
  excludeKeywords?: string[];
  /** @nullable */
  expectedSalaryMin?: number | null;
  /** @nullable */
  expectedSalaryMax?: number | null;
  isActive?: boolean;
}