# lib/api-zod/src/generated/types/sourceCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */

export interface SourceCount {
  source: string;
  count: number;
}