# lib/api-zod/src/generated/types/telegramStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QAHire – AI Job Curator API for Indian QA professionals
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramStatus {
  connected: boolean;
  /** @nullable */
  chatId?: string | null;
  /** @nullable */
  username?: string | null;
}